# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UnivMarket — Algerian University Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      parameters:
        - name: parent_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: active_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/popular:
    get:
      operationId: getPopularCategories
      tags: [categories]
      summary: Get popular categories with service counts
      parameters:
        - name: wilaya_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Popular categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryWithCount"

  # ─── SERVICES ───
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
...
  /disputes:
    get:
      operationId: listDisputes
      tags: [disputes]
      summary: List disputes (client, seller, or admin view)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Disputes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDisputes"

    post:
      operationId: openDispute
      tags: [disputes]
      summary: Open a dispute on an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "201":
          description: Dispute opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
  /disputes/{id}:
    get:
      operationId: getDispute
      tags: [disputes]
      summary: Get dispute details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispute detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
  /disputes/{id}/reply:
    post:
      operationId: replyToDispute
      tags: [disputes]
      summary: Seller replies to dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeReplyInput"
      responses:
        "200":
          description: Reply added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"

  # ─── NOTIFICATIONS ───
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, wilaya_id, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [client, seller, admin]
        wilaya_id:
          type: integer
        wilaya_name_fr:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        verified_at:
          type: ["string", "null"]
        banned_at:
          type: ["string", "null"]
        language_preference:
          type: string
          enum: [fr, ar, en]
          default: fr
        trust_score:
          type: ["number", "null"]
        referral_code:
          type: ["string", "null"]
        created_at:
          type: string
...
    Category:
      type: object
      required: [id, name_fr, slug, is_active]
      properties:
        id:
          type: integer
        name_fr:
          type: string
        name_ar:
          type: ["string", "null"]
        slug:
          type: string
        icon:
          type: ["string", "null"]
        parent_id:
          type: ["integer", "null"]
        sort_order:
          type: integer
        is_active:
          type: boolean
    CategoryWithCount:
      allOf:
        - $ref: "#/components/schemas/Category"
        - type: object
          required: [services_count]
          properties:
            services_count:
              type: integer

    # ─── SERVICE ───
    ServiceCard:
      type: object
      required: [id, title_fr, price, price_type, delivery_days, seller_id, seller_name, wilaya_id, wilaya_name_fr, category_id, status, created_at]
      properties:
        id:
          type: integer
        title_fr:
          type: string
        title_ar:
          type: ["string", "null"]
        price:
          type: number
        price_type:
          type: string
          enum: [fixed, hourly, negotiable]
        delivery_days:
          type: integer
        seller_id:
          type: integer
        seller_name:
          type: string
        seller_avatar:
          type: ["string", "null"]
        seller_verified:
          type: boolean
        seller_trust_score:
          type: ["number", "null"]
        wilaya_id:
          type: integer
        wilaya_name_fr:
          type: string
        category_id:
          type: integer
        category_name_fr:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        status:
          type: string
        is_featured:
          type: boolean
        avg_rating:
          type: ["number", "null"]
        reviews_count:
          type: integer
        views_count:
          type: integer
        created_at:
          type: string
...
    Order:
      type: object
      required: [id, client_id, service_id, seller_id, wilaya_id, status, total_price, commission_amount, payment_method, payment_status, created_at]
      properties:
        id:
          type: integer
        client_id:
          type: integer
        service_id:
          type: integer
        seller_id:
          type: integer
        wilaya_id:
          type: integer
        status:
          type: string
          enum: [pending, accepted, in_progress, delivered, completed, disputed, cancelled, refunded]
        total_price:
          type: number
        commission_amount:
          type: number
        payment_method:
          type: string
          enum: [paypal, cib, baridimob, dahabia]
        payment_status:
          type: string
          enum: [pending, confirmed, rejected, refunded]
        receipt_image:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        delivered_at:
          type: ["string", "null"]
        auto_complete_at:
          type: ["string", "null"]
        cancellation_reason:
          type: ["string", "null"]
        contract_signed_client:
          type: boolean
        contract_signed_seller:
          type: boolean
        created_at:
          type: string
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            service:
              $ref: "#/components/schemas/ServiceCard"
            client:
              $ref: "#/components/schemas/User"
            seller:
              $ref: "#/components/schemas/User"
            contract:
              $ref: "#/components/schemas/DigitalContract"
            escrow:
              $ref: "#/components/schemas/EscrowTransaction"
            review:
              $ref: "#/components/schemas/Review"
...
    DigitalContract:
      type: object
      required: [id, order_id, is_locked]
      properties:
        id:
          type: integer
        order_id:
          type: integer
        content_fr:
          type: ["string", "null"]
        client_signed_at:
          type: ["string", "null"]
        seller_signed_at:
          type: ["string", "null"]
        client_ip:
          type: ["string", "null"]
        seller_ip:
          type: ["string", "null"]
        pdf_url:
          type: ["string", "null"]
        qr_verification_code:
          type: ["string", "null"]
        is_locked:
          type: boolean
    EscrowTransaction:
      type: object
      required: [id, order_id, amount, commission_amount, status]
      properties:
        id:
          type: integer
        order_id:
          type: integer
        amount:
          type: number
        commission_amount:
          type: number
        status:
          type: string
          enum: [held, released, refunded, disputed, partially_refunded]
        held_at:
          type: ["string", "null"]
        released_at:
          type: ["string", "null"]
        auto_release_at:
          type: ["string", "null"]

    # ─── REVIEW ───
    Review:
      type: object
      required: [id, order_id, client_id, seller_id, rating, created_at]
      properties:
        id:
          type: integer
        order_id:
          type: integer
        client_id:
          type: integer
        seller_id:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        body:
          type: ["string", "null"]
        is_flagged:
          type: boolean
        seller_reply:
          type: ["string", "null"]
        reply_at:
          type: ["string", "null"]
        client_name:
          type: ["string", "null"]
        client_avatar:
          type: ["string", "null"]
        created_at:
          type: string
...
    Dispute:
      type: object
      required: [id, order_id, client_id, seller_id, reason, status, created_at]
      properties:
        id:
          type: integer
        order_id:
          type: integer
        client_id:
          type: integer
        seller_id:
          type: integer
        reason:
          type: string
        evidence_files:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [open, seller_replied, admin_review, resolved]
        admin_decision:
          type: ["string", "null"]
        refund_amount:
          type: ["number", "null"]
        resolved_at:
          type: ["string", "null"]
        created_at:
          type: string
    DisputeInput:
      type: object
      required: [order_id, reason]
      properties:
        order_id:
          type: integer
        reason:
          type: string
        evidence_files:
          type: array
          items:
            type: string
    DisputeReplyInput:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
        evidence_files:
          type: array
          items:
            type: string
...
    PaginatedDisputes:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Dispute"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── NOTIFICATIONS ───
...